Changed float to double, concerns over accuracy

+++ src/classifier/feed_forward.cpp
@@ -5,9 +5,9 @@
 
 FeedForward::FeedForward(){}
 
-float FeedForward::calculate_incoming_values(std::vector<Synapse>& incoming_synapses)
+double FeedForward::calculate_incoming_values(std::vector<Synapse>& incoming_synapses)
 {
-    float incoming_values = 0;
+    double incoming_values = 0;
     for(auto& synapse: incoming_synapses)
     {
         incoming_values += synapse.get_from_neuron_weighted_outgoing_value();
@@ -20,11 +20,11 @@ float FeedForward::calculate_incoming_values(std::vector<Synapse>& incoming_syna
  * The output will always range between 0 and 1, and usually closer to 0.5
  * Source - https://en.wikipedia.org/wiki/Logistic_function
  */
-float FeedForward::activate(float incoming_values)
+double FeedForward::activate(double incoming_values)
 {
     float euler_number = 2.71828;
-    float incoming_values_neg = -incoming_values;
-    float activation_result = 1/(1 + pow(euler_number, incoming_values_neg));
+    double incoming_values_neg = -incoming_values;
+    double activation_result = 1/(1 + pow(euler_number, incoming_values_neg));
     return activation_result;
 }
